+++ src/spr.hpp
@@ -127,6 +127,127 @@ array<T> get_SPR(vector<int> spawn_seasons, vector<int> fleet_regions, matrix<in
 }
 
 template <class T>
+array<T> get_SPR(vector<int> spawn_seasons, vector<int> fleet_regions, matrix<int> fleet_seasons, array<int> can_move, 
+  vector<int> mig_type, vector<T> fracyr_SSB, array<T> FAA, array<T> log_M, array<T> mu, vector<T> L, 
+  array<T> mature, array<T> waa_ssb, vector<T> fracyr_seasons, int age_specific, int bias_correct, array<T> log_NAA_sigma, int small_dim, int trace = 0, int numbers = 0){
+  /* 
+    calculate equilibrium spawning biomass per recruit (at age) by stock and region. If movement is set up approriately 
+    all fish can be made to return to a single spawning region for each stock.
+    NB: This version of get_SPR does not allow year indices. It is assumed that the yearly (or averaged) values are provided.
+       spawn_season: vector of indicators telling which season spawning occurs for each stock
+      fleet_regions: vector of indicators telling which region each fleet is operating
+      fleet_seasons: n_fleets x n_seasons; 0/1 indicating whether fleet is operating in the season
+           can_move: n_stocks x n_seasons x n_regions x n_regions: 0/1 determining whether movement can occur from one region to another
+           mig_type: n_stocks. 0 = migration after survival, 1 = movement and mortality simultaneous
+         fracyr_SSB: n_stocks:  size of interval from beginning of season to time of spawning within that season
+                FAA: fishing mortality: n_fleets x n_ages
+         log_M: log M (density-independent components): n_stocks x n_regions x n_ages
+                 mu: n_stocks x n_ages x n_seasons x n_regions x n_regions array of movement matrices
+                  L: n_regions. "extra" unobserved mortality
+             mature: n_stocks x n_ages proportion mature at age
+            waa_ssb: n_stocks x n_ages. weight at age
+     fracyr_seasons: n_seasons: length of intervals for each season
+       age_specific: 0/1 telling whether to return SSB/R by age or not. dimensions of returned arrays are different. 
+          small_dim: 0/1 telling whether the n_regions is "small." Different methods of inverting matrices.
+  */
+
+  int n_stocks = log_M.dim(0);
+  int n_regions = log_M.dim(1);
+  int n_ages = log_M.dim(2);
+  int n_fleets = FAA.rows();
+  int n_seasons = can_move.dim(1);
+  int P_dim = n_regions + n_fleets + 1;
+  if(trace) see("inside array<T> get_SPR");
+  array<T> SPRAA(n_stocks,n_ages,n_regions,n_regions); //SSB/R at age in each region column, given recruited in region row
+  SPRAA.setZero();
+  if(trace) see("T1");
+  matrix<T> I(P_dim,P_dim);
+  I.setZero();
+  for(int i = 0; i < P_dim; i++) I(i,i) = 1.0;
+  if(trace) see(I);
+  for(int s = 0; s < n_stocks; s++) {
+    matrix<T> cum_S_ya = get_S(I, n_regions); //cumulative survival up to age a
+    for(int a = 0; a < n_ages; a++) {
+      if(trace) see(a);
+      matrix<T> P_ya = I, P_spawn = I; //PTM for age and up to time of spawning
+      vector<T> M_t(n_regions);
+      for(int r = 0; r < n_regions; r++) M_t(r) = exp(log_M(s,r,a));
+      if(trace) see(M_t);
+      for(int t = 0; t < n_seasons; t++) {
+        if(trace) see(t);
+        if(trace) see(spawn_seasons(s)-1);
+        vector<T> F_t(n_fleets);
+        F_t.setZero();
+        for(int f = 0; f < n_fleets; f++) if(fleet_seasons(f,t)) F_t(f) = FAA(f,a);
+        if(trace) see(F_t);
+        matrix<T> mu_t(n_regions,n_regions);
+        mu_t.setZero();
+        matrix<int> can_move_t(n_regions, n_regions);
+        can_move_t.setZero();
+        if(n_regions>1) for(int r = 0; r < n_regions; r++) for(int rr = 0; rr < n_regions; rr++) {
+          can_move_t(r,rr) = can_move(s,t,r,rr);
+          mu_t(r,rr) = mu(s,a,t,r,rr);
+        }
+        if(trace) see(can_move_t);
+        if(trace) see(mu_t);
+        if(t == spawn_seasons(s)-1) {
+          if(trace) see(P_ya);
+          //P(0,t_spawn): PTM over entire year up to time of spawning
+          P_spawn = P_ya * get_P_t_base(fleet_regions, can_move_t, mig_type(s), fracyr_SSB(s), F_t, M_t, 
+            mu_t, L, trace);
+          if(trace) see(P_spawn);
+        }
+        //update PTM to end of season t P(0,s) * P(s,t) = P(0,t)
+        P_ya = P_ya * get_P_t_base(fleet_regions, can_move_t, mig_type(s), fracyr_seasons(t), F_t, M_t, 
+          mu_t, L, trace);
+        if(trace) see(P_ya);
+      }
+      // SSB/R at year and age = prob alive to up to age a x prob spawn at age a x waa x mature
+      //should be n_regions x n_regions
+      matrix<T> NPR_ya = cum_S_ya; //eq abundance at age a/recruit (at beginning of year y: Jan 1). Identity matrix for age 1 (0 in c++)
+      if(trace) see(NPR_ya);
+      
+      matrix<T> S_ya = get_S(P_ya, n_regions); //survival over year y for age a up to end of year y (or to age a+1 at beginning of year y+1)
+      if(bias_correct) for(int i = 0; i < n_regions; i++) for(int j = 0; j < n_regions; j++) {
+        //bias-correct for next age going to region j?
+        S_ya(i,j) = S_ya(i,j) * exp(-0.5*pow(exp(log_NAA_sigma(s,j,a+1)),2)); 
+      }
+      cum_S_ya = cum_S_ya * S_ya; //accumulate for next age
+      if(trace) see(cum_S_ya);
+      if(a == n_ages-1){ //plus group
+        // matrix<T> fundm(n_regions,n_regions);
+        matrix<T> fundm = -S_ya; //already has any bias correction by column.
+        if(small_dim) fundm = fundm.inverse(); else fundm = atomic::matinv(fundm);
+        //for plus group cum_S_ya = S_y,a-1 x (I - S_y,+)^-1
+        cum_S_ya = cum_S_ya * fundm;
+        NPR_ya = cum_S_ya; //change NPR_ya for plus group
+        if(trace) see(cum_S_ya);
+      }
+      //if numbers == 1 return numbers/recruit
+      matrix<T> SPR_ya = NPR_ya; 
+      //eq. SSB at age a (at time of spawning)/recruit
+      if(numbers==0) SPR_ya = SPR_ya * get_S(P_spawn, n_regions) * mature(s,a) * waa_ssb(s,a); 
+      if(trace) see(SPR_ya);
+      for(int i = 0; i < n_regions; i++) for(int j = 0; j < n_regions; j++) {
+        //SSB per Recruit in each region (should only have positive values in spawn_regions(s)-1?)
+        SPRAA(s,a,i,j) += SPR_ya(i,j); 
+      }
+    }
+  }
+  if(trace) see(SPRAA);
+  if(age_specific) {
+    return SPRAA; 
+  } else {//sum over ages
+    array<T> SPR(n_stocks,n_regions,n_regions);
+    SPR.setZero();
+    for(int s = 0; s < n_stocks; s++) for(int a = 0; a < n_ages; a++) for(int i = 0; i < n_regions; i++) for(int j = 0; j < n_regions; j++) {
+      SPR(s,i,j) += SPRAA(s,a,i,j);
+    }
+    return SPR;
+  }
+}
+
+template <class T>
 array<T> get_YPR_srf(vector<int> fleet_regions, matrix<int> fleet_seasons, array<int> can_move, 
   vector<int> mig_type, array<T> FAA, array<T> log_M, array<T> mu, vector<T> L, array<T> waacatch, 
   vector<T> fracyr_seasons, int age_specific, int small_dim){
@@ -207,9 +328,86 @@ array<T> get_YPR_srf(vector<int> fleet_regions, matrix<int> fleet_seasons, array
   }
 }
 
+template <class T>
+array<T> get_YPR_srf(vector<int> fleet_regions, matrix<int> fleet_seasons, array<int> can_move, 
+  vector<int> mig_type, array<T> FAA, array<T> log_M, array<T> mu, vector<T> L, array<T> waacatch, 
+  vector<T> fracyr_seasons, int age_specific, int bias_correct, array<T> log_NAA_sigma, int small_dim){
+  /* 
+    calculate equilibrium yield per recruit (at age) by stock and region.
+        fleet_regions: vector of indicators telling which region each fleet is operating
+        fleet_seasons: n_fleets x n_seasons; 0/1 indicating whether fleet is operating in the season
+             can_move: n_stocks x n_seasons x n_regions x n_regions: 0/1 determining whether movement can occur from one region to another
+             mig_type: n_stocks. 0 = migration after survival, 1 = movement and mortality simultaneous
+                  FAA: fishing mortality: n_fleets x n_ages
+                log_M: n_stocks x n_regions x n_ages
+                   mu: n_stocks x n_ages x n_seasons x n_regions x n_regions array of movement matrices
+                    L: n_regions. "extra" unobserved mortality
+             waacatch: n_fleets x n_ages. weight at age
+       fracyr_seasons: n_seasons: length of intervals for each season
+         age_specific: 0/1 telling whether to return SSB/R by age or not. dimensions of returned arrays are different. 
+            small_dim: 0/1 telling whether the n_regions is "small." Different methods of inverting matrices.
+  */
+
+  int n_stocks = log_M.dim(0);
+  int n_regions = log_M.dim(1);
+  int n_ages = log_M.dim(2);
+  int n_fleets = FAA.rows();
+  int n_seasons = can_move.dim(1);
+  int P_dim = n_regions + n_fleets + 1;
+  array<T> YPRAA(n_stocks,n_ages,n_regions,n_fleets); //Yield/R at age in each fleet column, given recruited in region row
+  YPRAA.setZero();
+  matrix<T> W(n_fleets,n_fleets);
+  W.setZero();
+
+  matrix<T> I(P_dim,P_dim);
+  I.setZero();
+  for(int i = 0; i < P_dim; i++) I(i,i) = 1.0;
+
+  for(int s = 0; s < n_stocks; s++) {
+    matrix<T> cum_S_ya = get_S(I, n_regions); //cumulative survival up to age a
+    for(int a = 0; a < n_ages; a++) {
+      for(int i = 0; i < n_fleets; i++) W(i,i) = waacatch(i,a);
+      matrix<T> P_ya = I; //PTM for year and age
+      for(int t = 0; t < n_seasons; t++) {
+        //update PTM to end of season t P(0,s) * P(s,t) = P(0,t)
+        P_ya = P_ya * get_P_t(a, s, t, fleet_regions, fleet_seasons, can_move, mig_type, fracyr_seasons(t), FAA.matrix(), log_M, mu, L);
+      }
+      // N/recruit at beginning of year (Jan 1) y at age a. Identity matrix for age 1 (0 in c++)
+      matrix<T> NPR_ya = cum_S_ya; 
+      // matrix<T> YPR_ya = cum_S_ya * get_D(P_ya, n_regions,n_fleets) * W; //should be n_regions x n_fleets
+      matrix<T> S_ya = get_S(P_ya, n_regions); //survival over year y for age a up to end of year y (or to age a+1 at beginning of year y+1)
+      if(bias_correct) for(int i = 0; i < n_regions; i++) for(int j = 0; j < n_regions; j++) {
+        //bias-correct for next age going to region j?
+        S_ya(i,j) = S_ya(i,j) * exp(-0.5*pow(exp(log_NAA_sigma(s,j,a+1)),2)); 
+      }
+      cum_S_ya = cum_S_ya * S_ya; //accumulate for next age
+      if(a == n_ages-1){ //plus group
+        // matrix<T> fundm(n_regions,n_regions);
+        matrix<T> fundm = -S_ya; //already has any bias correction by column.
+        if(small_dim) fundm = fundm.inverse(); else fundm = atomic::matinv(fundm);
+        //for plus group cum_S_ya = S_y,a-1 x (I - S_y,+)^-1
+        cum_S_ya = cum_S_ya * fundm;
+        NPR_ya = cum_S_ya; //revise N/recruit on jan 1 for plus group
+      }
+      matrix<T> YPR_ya = NPR_ya * get_D(P_ya, n_regions,n_fleets) * W; //(n_regions x n_fleets) yield/recruit starting year in region r and being caught in fleet f.
+      for(int i = 0; i < n_regions; i++) for(int j = 0; j < n_fleets; j++) YPRAA(s,a,i,j) = YPR_ya(i,j);
+    }
+  }
+  if(age_specific) {
+    return YPRAA; 
+  } else { //sum across age
+    array<T> YPR(n_stocks,n_regions,n_fleets);
+    YPR.setZero();
+    for(int s = 0; s < n_stocks; s++) for(int a = 0; a < n_ages; a++) for(int i = 0; i < n_regions; i++) for(int j = 0; j < n_fleets; j++) {
+      YPR(s,i,j) += YPRAA(s,a,i,j);
+    }
+    return YPR;
+  }
+}
+
 template <class Type>
 matrix<Type> get_RXSPR(array<Type> all_NAA, vector<int> spawn_regions, int n_years_model, int n_years_proj, 
-  int XSPR_R_opt, vector<int> XSPR_R_avg_yrs){
+  int XSPR_R_opt, vector<int> XSPR_R_avg_yrs, array<Type> log_NAA_sigma){
   //R_XSPR is needed for projections and reference points
   array<Type> NAA = extract_NAA(all_NAA);
   array<Type> pred_NAA = extract_pred_NAA(all_NAA);
@@ -225,7 +423,7 @@ matrix<Type> get_RXSPR(array<Type> all_NAA, vector<int> spawn_regions, int n_yea
       for(int y = 0; y < XSPR_R_avg_yrs.size(); y++) {
         if(XSPR_R_opt == 2) avg_R(s) += NAA(s,spawn_regions(s)-1,XSPR_R_avg_yrs(y),0);
         if(XSPR_R_opt == 4) avg_R(s) += pred_NAA(s,spawn_regions(s)-1,XSPR_R_avg_yrs(y),0);
-        if(XSPR_R_opt == 5) avg_R(s) += pred_NAA(s,spawn_regions(s)-1,XSPR_R_avg_yrs(y),0);
+        if(XSPR_R_opt == 5) avg_R(s) += pred_NAA(s,spawn_regions(s)-1,XSPR_R_avg_yrs(y),0) * exp(-0.5*pow(exp(log_NAA_sigma(s,spawn_regions(s)-1,0)),2));
       }
       avg_R(s) /= Type(XSPR_R_avg_yrs.size());
       for(int y = 0; y < n_years_model + n_years_proj; y++) R_XSPR(y,s) = avg_R(s);
@@ -255,6 +453,8 @@ struct spr_F_spatial {
   vector<Type> fracyr_seasons;
   vector<Type> SPR_weights; //how to weight stock-specific SSB/R for aggregate SSB/R.
   int age_specific; 
+  int bias_correct;
+  array<Type> log_NAA_sigma;
   int small_dim;
   int trace=0;
 
@@ -276,6 +476,8 @@ struct spr_F_spatial {
   vector<Type> fracyr_seasons_,
   vector<Type> SPR_weights_,
   int age_specific_, 
+  int bias_correct_,
+  array<Type> log_NAA_sigma_,
   int small_dim_,
   int trace_) :
     spawn_seasons(spawn_seasons_), 
@@ -294,6 +496,8 @@ struct spr_F_spatial {
     fracyr_seasons(fracyr_seasons_),
     SPR_weights(SPR_weights_),
     age_specific(age_specific_),
+    bias_correct(bias_correct_),
+    log_NAA_sigma(log_NAA_sigma_),
     small_dim(small_dim_), 
     trace(trace_) {}
 
@@ -321,9 +525,10 @@ struct spr_F_spatial {
     }
     if(trace) see(FAA);
     vector<T> fracyrssbT = fracyr_SSB.template cast<T>();
-    array<T> logMbaseT(log_M.dim(0),log_M.dim(1),log_M.dim(2));
+    array<T> logMbaseT(n_stocks,n_regions,n_ages), log_NAA_sigmaT(n_stocks, n_regions, n_ages);
     for(int s = 0; s < n_stocks; s++) for(int r = 0; r < n_regions; r++) for(int a = 0; a < n_ages; a++){
       logMbaseT(s,r,a) = T(log_M(s,r,a));
+      log_NAA_sigmaT(s,r,a) = T(log_NAA_sigma(s,r,a));
     }
     if(trace) see(logMbaseT.dim);
     if(trace) see(mu.dim);
@@ -357,7 +562,10 @@ struct spr_F_spatial {
       matT, 
       waassbT, 
       fracyrseasonT, 
-      0, small_dim, 0, 0);
+      0, 
+      bias_correct,
+      log_NAA_sigmaT,
+      small_dim, 0, 0);
     if(trace) see("end get_SPR spr_F_spatial");
 
     //weighted-average of SSB/R returned
@@ -389,7 +597,10 @@ struct spr_F_spatial {
         matT, 
         waassbT, 
         fracyrseasonT, 
-        0, small_dim, 0, 0);
+        0, 
+        bias_correct,
+        log_NAA_sigmaT,
+        small_dim, 0, 0);
       T SPR0 = 0;
       for(int s = 0; s < n_stocks; s++) SPR0 += T(SPR_weights(s)) * SPR0_sr(s,spawn_regions(s)-1,spawn_regions(s)-1); 
       if(trace) see(SPR0);
@@ -405,7 +616,8 @@ template <class Type>
 vector<Type> get_FXSPR(vector<int> spawn_seasons, vector<int> spawn_regions, vector<int> fleet_regions, matrix<int> fleet_seasons,
   array<int> can_move, vector<int> mig_type, vector<Type> ssbfrac, array<Type> sel, array<Type> log_M, array<Type> mu, 
   vector<Type> L, array<Type> mat,  array<Type> waassb, vector<Type> fracyr_seasons, vector<Type> R_XSPR, 
-  Type percentSPR, vector<Type> SPR_weights, int SPR_weight_type, int small_dim, Type F_init, int n_iter, int trace, int by_region = 0) {
+  Type percentSPR, vector<Type> SPR_weights, int SPR_weight_type, int bias_correct, array<Type> log_NAA_sigma, 
+  int small_dim, Type F_init, int n_iter, int trace, int by_region = 0) {
   int n_stocks = spawn_seasons.size();
   int n_fleets = fleet_regions.size();
   int n_ages = mat.cols();
@@ -419,7 +631,7 @@ vector<Type> get_FXSPR(vector<int> spawn_seasons, vector<int> spawn_regions, vec
   array<Type> FAA0(n_fleets,n_ages);
   FAA0.setZero();
   array<Type> SPR0_all = get_SPR(spawn_seasons, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, FAA0, log_M, mu, L, 
-    mat, waassb, fracyr_seasons, 0, small_dim, 0, 0);
+    mat, waassb, fracyr_seasons, 0, bias_correct, log_NAA_sigma, small_dim, 0, 0);
   if(trace) see(SPR0_all);
   int n_F = 1;
   if(by_region) n_F = n_regions;
@@ -439,7 +651,7 @@ vector<Type> get_FXSPR(vector<int> spawn_seasons, vector<int> spawn_regions, vec
   for(int r = 0; r < n_F; r++) log_FXSPR_iter(0,r) = log(F_init);
   if(trace) see(log_FXSPR_iter.row(0));
   spr_F_spatial<Type> sprF(spawn_seasons, spawn_regions, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, sel, log_M,
-    mu, L, mat, waassb, fracyr_seasons, SPR_weights, 0, small_dim, trace);    
+    mu, L, mat, waassb, fracyr_seasons, SPR_weights, 0, bias_correct, log_NAA_sigma, small_dim, trace);    
   if(trace) see("after spr_F_spatial sprF defined");
   //trace = 0;
   for(int i=0; i<n_iter-1; i++) {
@@ -491,7 +703,8 @@ template <class Type>
 vector<Type> get_FXSPR(vector<int> spawn_seasons, vector<int> spawn_regions, vector<int> fleet_regions, matrix<int> fleet_seasons,
   array<int> can_move, vector<int> mig_type, vector<Type> ssbfrac, array<Type> sel, array<Type> log_M, array<Type> mu, 
   vector<Type> L, array<Type> mat,  array<Type> waassb, vector<Type> fracyr_seasons, vector<Type> R_XSPR, vector<Type> log_SPR0,
-  Type percentSPR, vector<Type> SPR_weights, int SPR_weight_type, int small_dim, Type F_init, int n_iter, int trace, int by_region = 0) {
+  Type percentSPR, vector<Type> SPR_weights, int SPR_weight_type, int bias_correct, array<Type> log_NAA_sigma, 
+  int small_dim, Type F_init, int n_iter, int trace, int by_region = 0) {
   int n_stocks = spawn_seasons.size();
   int n_fleets = fleet_regions.size();
   int n_ages = mat.cols();
@@ -520,7 +733,7 @@ vector<Type> get_FXSPR(vector<int> spawn_seasons, vector<int> spawn_regions, vec
   for(int r = 0; r < n_F; r++) log_FXSPR_iter(0,r) = log(F_init);
   if(trace) see(log_FXSPR_iter.row(0));
   spr_F_spatial<Type> sprF(spawn_seasons, spawn_regions, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, sel, log_M,
-    mu, L, mat, waassb, fracyr_seasons, SPR_weights, 0, small_dim, trace);    
+    mu, L, mat, waassb, fracyr_seasons, SPR_weights, 0, bias_correct, log_NAA_sigma, small_dim, trace);    
   if(trace) see("after spr_F_spatial sprF defined");
   //trace = 0;
   for(int i=0; i<n_iter-1; i++) {
@@ -558,7 +771,7 @@ vector< array <Type> > get_SPR_res(vector<Type> SPR_weights, array<Type> log_M, 
   array<Type> mature, Type percentSPR, array<Type> NAA, matrix<Type> fracyr_SSB, Type F_init, 
   vector<int> years_M, vector<int> years_mu, vector<int> years_L, vector<int> years_mat, vector<int> years_sel, 
   vector<int> years_waa_ssb, vector<int> years_waa_catch, vector<Type> R_XSPR,
-  int small_dim, int SPR_weight_type, int trace = 0, int n_iter = 10) {
+  int small_dim, int SPR_weight_type, int bias_correct, array<Type> log_NAA_sigma, int trace = 0, int n_iter = 10) {
   //gets SPR-based BRP information for a year, or inputs may be averaged over specified years.  
   if(trace) see("inside get_SPR_res");
   //see(SPR_weights);
@@ -611,9 +824,9 @@ vector< array <Type> > get_SPR_res(vector<Type> SPR_weights, array<Type> log_M, 
   FAA0.setZero();
   //equil abundance/R (Jan 1)
   array<Type> NAAPR0_all = get_SPR(spawn_seasons, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, FAA0, log_M_avg, mu_avg, L_avg, 
-    mat, waa_ssb_avg, fracyr_seasons, 1, small_dim, 0, 1); 
+    mat, waa_ssb_avg, fracyr_seasons, 1, bias_correct, log_NAA_sigma, small_dim, 0, 1); 
   array<Type> SPRAA0_all = get_SPR(spawn_seasons, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, FAA0, log_M_avg, mu_avg, L_avg, 
-    mat, waa_ssb_avg, fracyr_seasons, 1, small_dim, 0, 0);
+    mat, waa_ssb_avg, fracyr_seasons, 1, bias_correct, log_NAA_sigma, small_dim, 0, 0);
   if(trace) see(SPRAA0_all.dim);
   array<Type> SPR0_all(n_stocks, n_regions,n_regions);
   SPR0_all.setZero();
@@ -633,7 +846,7 @@ vector< array <Type> > get_SPR_res(vector<Type> SPR_weights, array<Type> log_M, 
   if(trace) see(log_FXSPR_iter(0));
   // trace = 0;
   spr_F_spatial<Type> sprF(spawn_seasons, spawn_regions, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, sel, log_M_avg,
-    mu_avg, L_avg, mat, waa_ssb_avg, fracyr_seasons, SPR_weights, 0, small_dim, trace);
+    mu_avg, L_avg, mat, waa_ssb_avg, fracyr_seasons, SPR_weights, 0, bias_correct, log_NAA_sigma, small_dim, trace);
   if(trace) see("after spr_F_spatial sprF defined");
   // trace = 1;
   for(int i=0; i<n_iter-1; i++) {
@@ -663,9 +876,9 @@ vector< array <Type> > get_SPR_res(vector<Type> SPR_weights, array<Type> log_M, 
   if(trace) see(log_FAA_XSPR);
   //eq NAA/R at FXSPR
   array<Type> NAAPR_FXSPR_all = get_SPR(spawn_seasons, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, FAA_XSPR, log_M_avg, mu_avg, L_avg, 
-    mat, waa_ssb_avg, fracyr_seasons, 1, small_dim, 0, 1);
+    mat, waa_ssb_avg, fracyr_seasons, 1, bias_correct, log_NAA_sigma, small_dim, 0, 1);
   array<Type> SPRAA_all = get_SPR(spawn_seasons, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, FAA_XSPR, log_M_avg, mu_avg, L_avg, 
-    mat, waa_ssb_avg, fracyr_seasons, 1, small_dim, 0, 0);
+    mat, waa_ssb_avg, fracyr_seasons, 1, bias_correct, log_NAA_sigma, small_dim, 0, 0);
   if(trace) see(SPRAA_all.dim);
   array<Type> SPR_all(n_stocks, n_regions,n_regions);
   SPR_all.setZero();
@@ -674,7 +887,7 @@ vector< array <Type> > get_SPR_res(vector<Type> SPR_weights, array<Type> log_M, 
   }
   if(trace) see(SPR_all);
   array<Type> YPR_srf = get_YPR_srf(fleet_regions, fleet_seasons, can_move, mig_type, FAA_XSPR, log_M_avg, mu_avg, L_avg, waa_catch_avg, 
-    fracyr_seasons, 0, small_dim); //n_stocks x n_regions x n_fleets (should be 0 for regions not fleet_regions(f)-1)
+    fracyr_seasons, 0, bias_correct, log_NAA_sigma, small_dim); //n_stocks x n_regions x n_fleets (should be 0 for regions not fleet_regions(f)-1)
   //for each stock/fleet and also the weighted average/total
   if(trace) see(YPR_srf);
   array<Type> log_SPR(1,n_stocks + 1), log_SPR0(1,n_stocks+1),log_Y_XSPR(1,n_fleets+n_regions+1), log_SSB_XSPR(1,n_stocks+1), log_YPR_XSPR(n_stocks,n_fleets+1); 
@@ -755,7 +968,10 @@ vector< array <Type> > get_annual_SPR_res(vector<Type> SPR_weights, array<Type> 
   vector<int> which_F_age, array<Type> waa_ssb, array<Type> waa_catch,
   array<Type> mature, Type percentSPR, array<Type> NAA, matrix<Type> fracyr_SSB, vector<Type> F_init,  
   matrix<Type> R_XSPR,
-  int small_dim, int SPR_weight_type, int trace = 0, int n_iter = 10){
+  int small_dim, int SPR_weight_type, 
+  int bias_correct,
+  array<Type> log_NAA_sigma,
+  int trace = 0, int n_iter = 10){
   int ny = which_F_age.size();
   int n_fleets = waa_catch.dim(0);
   int n_regions = can_move.dim(2);
@@ -777,7 +993,7 @@ vector< array <Type> > get_annual_SPR_res(vector<Type> SPR_weights, array<Type> 
     vector< array<Type>> SPR_res_y = get_SPR_res(SPR_weights, log_M, FAA, spawn_seasons,  spawn_regions, fleet_regions, 
       fleet_seasons, fracyr_seasons, can_move, must_move, mig_type, trans_mu_base, L, which_F_age(y), 
       waa_ssb, waa_catch, mature, percentSPR, NAA, fracyr_SSB, F_init(y), yvec, yvec, yvec, yvec, yvec, yvec, yvec, 
-      vector<Type> (R_XSPR.row(y)), small_dim, SPR_weight_type, trace = trace, n_iter = n_iter);
+      vector<Type> (R_XSPR.row(y)), small_dim, SPR_weight_type, bias_correct, log_NAA_sigma, trace = trace, n_iter = n_iter);
     for(int f = 0; f <= n_fleets+n_regions; f++) for(int a = 0; a < n_ages; a++){
       log_FAA_XSPR(f,y,a) = SPR_res_y(0)(f,a);
     }
@@ -813,6 +1029,8 @@ array <Type> get_annual_SPR0_at_age(array<Type> log_M, vector<int> spawn_seasons
   matrix<Type> L,
   array<Type> waa_ssb, 
   array<Type> mature, matrix<Type> fracyr_SSB,
+  int bias_correct,
+  array<Type> log_NAA_sigma,
   int small_dim, int trace = 0){
   
   int ny = log_M.dim(2);
@@ -854,7 +1072,7 @@ array <Type> get_annual_SPR0_at_age(array<Type> log_M, vector<int> spawn_seasons
     // see(mu_avg.dim);
 
     array<Type> SPR0AA_y = get_SPR(spawn_seasons, fleet_regions, fleet_seasons, can_move, mig_type, ssbfrac, FAA0, log_M_avg, mu_avg, L_avg, 
-      mat, waa_ssb_avg, fracyr_seasons, 1, small_dim, 0);
+      mat, waa_ssb_avg, fracyr_seasons, 1, bias_correct, log_NAA_sigma, small_dim, 0);
     // see(SPR0AA_y.dim);
     for(int s = 0; s < n_stocks; s++) for(int a = 0; a < n_ages; a++)for(int r = 0; r < n_regions; r++) for(int rr = 0; rr < n_regions; rr++){
       // see(s);